add new table for word spellings, refactor schema to use word_ids
had to change several generator comprehensions to list comprehensions because sqlite does not allow
cursors to be used in recursive calls. and generators are recursive?
I don't anticipate this being a major issue, as the input to the generators originated from the JSON,
which itself had to be all in memory at one point.
meaning: the instantaneous memory usage is not asymptotically greater than before